requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.4
django-allauth==0.33.0
django-filter==1.0.4
django-grappelli==2.10.1
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
djangorestframework==3.6.4
drf-extensions==0.3.1
dry-rest-permissions==0.1.9
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
MySQL-python==1.2.5
oauthlib==2.0.2
openapi-codec==1.3.2
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.10.0
static3==0.7.0
uritemplate==3.0.0
urllib3==1.22
